# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API GIMA",
        "description": "Documentación de la API para el proyecto GIMA (Gestión de Inventario y Mantenimiento)",
        "contact": {
            "email": "admin@gima.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor API Principal"
        }
    ],
    "paths": {
...
        "/api/inventario/proveedores": {
            "get": {
                "tags": [
                    "Inventario - Proveedores"
                ],
                "summary": "Listar proveedores",
                "operationId": "301f17217141277d6e740fa974fd5c02",
                "responses": {
                    "200": {
                        "description": "Lista de proveedores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Proveedor"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inventario - Proveedores"
                ],
                "summary": "Crear proveedor",
                "operationId": "fc0a610348e5a2dee5ab929915071eca",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "contacto": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proveedor creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proveedor"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inventario/proveedores/{id}": {
            "get": {
                "tags": [
                    "Inventario - Proveedores"
                ],
                "summary": "Ver proveedor",
                "operationId": "7daff0577c14c6361af65bf135da79d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proveedor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proveedor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Inventario - Proveedores"
                ],
                "summary": "Actualizar proveedor",
                "operationId": "0a057f35c69819a03db0377485edfb62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proveedor actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proveedor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inventario - Proveedores"
                ],
                "summary": "Eliminar proveedor",
                "operationId": "2c82128d45270644170efa568b8ae1fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Proveedor": {
                "title": "Proveedor",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "contacto": {
                        "type": "string",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}